+++ backend/apid/graphql/check_test.go
@@ -1,0 +1,49 @@
+package graphql
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/sensu/sensu-go/backend/apid/graphql/schema"
+	"github.com/sensu/sensu-go/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestCheckTypeHistoryFieldImpl(t *testing.T) {
+	testCases := []struct {
+		expectedLen int
+		firstArg    int
+	}{
+		{
+			expectedLen: 21,
+			firstArg:    50,
+		},
+		{
+			expectedLen: 10,
+			firstArg:    10,
+		},
+		{
+			expectedLen: 0,
+			firstArg:    0,
+		},
+		{
+			expectedLen: 0,
+			firstArg:    -10,
+		},
+	}
+
+	check := types.FixtureCheck("test")
+	for _, tc := range testCases {
+		t.Run(fmt.Sprintf("w/ argument of %d", tc.expectedLen), func(t *testing.T) {
+			params := schema.CheckHistoryFieldResolverParams{}
+			params.Source = check
+			params.Args.First = tc.firstArg
+
+			impl := checkImpl{}
+			res, err := impl.History(params)
+			require.NoError(t, err)
+			assert.Len(t, res, tc.expectedLen)
+		})
+	}
+}