+++ dashboard/src/components/ResetStyles.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { withStyles } from "material-ui/styles";
+
+import reset from "../reset.css";
+
+class ResetStyles extends React.PureComponent {
+  render() {
+    return null;
+  }
+}
+
+export default withStyles({ "@global": reset })(ResetStyles);