+++ CHANGELOG.md
@@ -19,8 +19,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added extension registry to the store.
 - Add sensuctl create command.
 - Adds a statsd server to the sensu-agent which runs statsd at a configurable
-flush interval and converts gostatsd metrics to Sensu Metric Format. These
-metric events are swallowed for the time being.
+flush interval and converts gostatsd metrics to Sensu Metric Format.
 - Add event filtering to extensions.
 - Proper 404 page for web UI.
 - Add sensuctl extension command.
@@ -29,6 +28,11 @@ metric events are swallowed for the time being.
 - Add the `--statsd-event-handlers` flag to sensu-agent which configures the
 event handlers for statsd metrics.
 - Add default user with username "sensu" with global, read-only permissions.
+- Add end-to-end test for extensions.
+- Add configuration setting for backend and agent log level.
+- Add extension package for building third-party Sensu extensions in Go.
+- Add the `--statsd-disable` flag to sensu-agent which configures the
+statsd listener. The listener is enabled by default.
 
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024
@@ -49,6 +53,8 @@ and yarn are now dependencies for building the backend.
 - Don't allow the bus to drop messages.
 - Events list can properly be viewed on mobile.
 - Updated Sirupsen/logrus to sirupsen/logrus and other applicable dependencies using the former.
+- Set default log level to 'warn'.
+- Optimize check marshaling.
 
 ### Fixed
 - Terminate processes gracefully in e2e tests, allowing ports to be reused.
@@ -61,6 +67,9 @@ and yarn are now dependencies for building the backend.
 - Properly log the event being handled in pipelined
 - The http_check.sh example script now hides its output
 - Silenced entries using an asterisk can be deleted
+- Events created from the metrics passed to the statsd listener are no longer
+swallowed. The events are sent through the pipeline.
+- Fixed a bug where the Issued field was never populated.
 
 ### Added
 - Support for managing mutators via sensuctl.